fixed instanciated stations to not be selectable. Ingame Trainstation Creation, scaling. Fix for Instanced Trainstation in Game scene.
added old railwaymap to check lines to add in the very empty parts.
Added possibility to create a new named station at the mouse position with the home/pos1 key. (The name is taken from the inputField that is currently part of the HUD in Game.unity scene.)
(some lines have been already added and moved accordingly for the final map)

Was unable to create Trainlines when the TrainStations were children of the TrainStationManager. So rewrote the saving (well the updating the data to be saved part) code to iterate over all the gameobjects (since I can't just conveniently iterate over the children of TrainStationManager anymore)

the resulting saved csv is perfectly fine. It's just in reverse order so Starting with Sommerfeld and ending with Aachen.